+++ tests/Unit/Storage/Factory/DbAdapterStorageFactoryTest.php
@@ -77,4 +77,24 @@ final class DbAdapterStorageFactoryTest extends TestCase
         $this->expectException(RuntimeException::class);
         (new DbAdapterStorageFactory())($container);
     }
+
+    public function testThrowsWhenContextMapIsNotStringToString(): void
+    {
+        $container = new ArrayContainer([
+            'config'     => [
+                'log' => [
+                    'storage' => [
+                        'options' => [
+                            'adapter' => 'db.adapter',
+                            'context' => ['student' => ['nested']],
+                        ],
+                    ],
+                ],
+            ],
+            'db.adapter' => $this->createMock(AdapterInterface::class),
+        ]);
+
+        $this->expectException(RuntimeException::class);
+        (new DbAdapterStorageFactory())($container);
+    }
 }